gemini_chatbot/requirements.txt
gradio
fastapi
uvicorn
python-dotenv
google-generativeai
